# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业级任务管理系统的 RESTful API 接口文档",
        "title": "任务管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "192.168.12.48:8888",
    "basePath": "/api/v1",
    "paths": {
...
        "/review-sessions/{sessionId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "查看审核会话的详细信息和投票情况",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "获取审核会话详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "审核会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewSessionResponse"
                        }
                    }
                }
            }
        },
        "/review-sessions/{sessionId}/finalize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建人做出最终审核决策",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "最终决策",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "审核会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "决策",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FinalizeReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "决策完成",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/review-sessions/{sessionId}/invite-jury": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建人将单人审核转为陪审团审核",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "邀请陪审团",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "审核会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "陪审团成员",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InviteJuryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "邀请成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/review-sessions/{sessionId}/jury/{juryMemberId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建人移除指定的陪审团成员",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "移除陪审团成员",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "审核会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "陪审团成员ID",
                        "name": "juryMemberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "移除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/review-sessions/{sessionId}/opinion": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "陪审团成员提交审核意见",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "提交审核意见",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "审核会话ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审核意见",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitReviewOpinionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "执行人接受被指派的任务，需求类任务接受后进入待提交方案状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "接受任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "接受成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/execution-plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "执行人提交需求任务的执行计划和目标（合并提交）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "提交执行计划和目标",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "执行计划和目标",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitExecutionPlanWithGoalsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/goals": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "此接口已废弃，请使用 POST /tasks/{id}/solution 提交方案",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "提交目标和方案（已废弃）",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "目标和方案",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitGoalsAndSolutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/plan": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "此接口已废弃，请使用 POST /tasks/{id}/execution-plan 同时提交执行计划和目标",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "提交执行计划（已废弃）",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "执行计划",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitExecutionPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "执行人拒绝被指派的任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "拒绝任务",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "拒绝原因",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RejectTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "拒绝成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/tasks/{id}/review": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建人发起方案或执行计划审核",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "发起审核",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审核配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.InitiateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "审核发起成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/tasks/{id}/solution": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "执行人提交需求任务的解决方案（不包含目标，目标将在执行计划阶段提交）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务流程"
                ],
                "summary": "提交解决方案",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "任务ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "解决方案",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitSolutionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.FinalizeReviewRequest": {
            "type": "object",
            "properties": {
                "approved": {
                    "description": "是否批准（true=批准, false=拒绝）",
                    "type": "boolean"
                },
                "comment": {
                    "description": "最终决策备注",
                    "type": "string"
                }
            }
        },
...
        "dto.InviteJuryRequest": {
            "type": "object",
            "required": [
                "jury_member_ids",
                "required_approvals"
            ],
            "properties": {
                "jury_member_ids": {
                    "description": "陪审团成员ID列表（要邀请为陪审团成员的用户）",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "integer"
                    }
                },
                "required_approvals": {
                    "description": "所需批准数（陪审团审核需要的最少批准数）",
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
        "dto.ReviewSessionResponse": {
            "type": "object",
            "properties": {
                "final_decision": {
                    "description": "最终决议（可选）",
                    "type": "string"
                },
                "final_decision_comment": {
                    "description": "最终决议备注",
                    "type": "string"
                },
                "id": {
                    "description": "审核会话ID",
                    "type": "integer"
                },
                "required_approvals": {
                    "description": "所需批准数",
                    "type": "integer"
                },
                "review_mode": {
                    "description": "审核模式（single=单人审核, jury=陪审团审核）",
                    "type": "string"
                },
                "review_records": {
                    "description": "审核记录列表（各个审核人的意见，可选）",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReviewRecordResponse"
                    }
                },
                "review_type": {
                    "description": "审核类型（goal_review=目标审核, solution_review=方案审核, plan_review=计划审核）",
                    "type": "string"
                },
                "status": {
                    "description": "审核状态（pending=进行中, approved=已批准, rejected=已拒绝）",
                    "type": "string"
                },
                "target_id": {
                    "description": "被审核的对象ID",
                    "type": "integer"
                },
                "target_type": {
                    "description": "目标类型",
                    "type": "string"
                },
                "task_id": {
                    "description": "关联的任务ID",
                    "type": "integer"
                }
            }
        },
...
        "dto.SubmitExecutionPlanWithGoalsRequest": {
            "type": "object",
            "required": [
                "goals",
                "implementation_steps",
                "tech_stack"
            ],
            "properties": {
                "goals": {
                    "description": "目标列表（至少一个目标）",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.GoalItem"
                    }
                },
                "implementation_steps": {
                    "description": "实施步骤（具体的实现步骤，JSON格式）",
                    "type": "object",
                    "additionalProperties": true
                },
                "resource_requirements": {
                    "description": "资源需求（所需资源、人力、设备等）",
                    "type": "string"
                },
                "risk_assessment": {
                    "description": "风险评估（可能的风险和应对方案）",
                    "type": "string"
                },
                "tech_stack": {
                    "description": "技术栈（使用的技术和工具）",
                    "type": "string"
                }
            }
        },
        "dto.SubmitGoalsAndSolutionRequest": {
            "type": "object",
            "required": [
                "goals",
                "solution"
            ],
            "properties": {
                "goals": {
                    "description": "目标列表",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/dto.GoalItem"
                    }
                },
                "solution": {
                    "description": "方案内容",
                    "allOf": [
                        {
                            "$ref": "#/definitions/dto.SolutionItem"
                        }
                    ]
                }
            }
        },
        "dto.SubmitReviewOpinionRequest": {
            "type": "object",
            "required": [
                "opinion"
            ],
            "properties": {
                "comment": {
                    "description": "审核备注（审核人的意见和建议）",
                    "type": "string"
                },
                "opinion": {
                    "description": "审核意见（approve=批准, reject=拒绝, abstain=弃权）",
                    "type": "string",
                    "enum": [
                        "approve",
                        "reject",
                        "abstain"
                    ]
                },
                "score": {
                    "description": "评分（可选，用于量化评估）",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}